requirements.txt
fastapi==0.115.9
websockets==11.0.3
uvicorn==0.29.0
openai==1.19.0
chromadb==1.0.17
json_repair==0.29.2
Jinja2==3.1.3
pypinyin==0.51.0
python-jose==3.3.0
PyMySQL==1.1.1
zhipuai==2.1.5.20250106
SQLAlchemy==2.0.29
XdbSearchIP==1.0.2
sqlalchemy-history==2.1.4
xlrd==2.0.1
openpyxl==3.1.2
pandas==2.2.2
joblib==1.4.0
python-multipart==0.0.9
httpx==0.27.0
chardet==5.2.0
pypdfium2==4.30.0
numexpr==2.11.0
httpx_sse==0.4.1
sseclient==0.0.27
pydantic_extra_types==2.10.5
python-docx==1.2.0
bs4==0.0.2
pyzipper==0.3.6
fastapi-offline==1.7.3